Change weight tensor to float32

+++ main.py
@@ -121,7 +121,7 @@ def run_experiment(cfg: OmegaConf) -> None:
 
     params_to_update = model.parameters()
     optimizer = optim.Adam(params=params_to_update, lr=lr)
-    criterion = nn.CrossEntropyLoss(weight=torch.tensor(class_weights).to(DEVICE))
+    criterion = nn.CrossEntropyLoss(weight=torch.tensor(class_weights, dtype=torch.float32).to(DEVICE))
 
     trained_model = train_model(model, dataloaders, criterion, optimizer, DEVICE, num_epochs, print_batch)
